+++ jax/_src/numpy/array_methods.py
@@ -23,7 +23,6 @@ __all__ = ['register_jax_array_methods']
 
 import abc
 from functools import partial, wraps
-import inspect
 from typing import Any, Optional, Union
 import warnings
 
@@ -465,27 +464,6 @@ class _IndexUpdateHelper:
     return f"_IndexUpdateHelper({repr(self.array)})"
 
 
-# TODO(jakevdp): remove these deprecation warnings after June 2023
-def allow_pass_by_position_with_warning(f):
-  @wraps(f)
-  def wrapped(*args, **kwargs):
-    sig = inspect.signature(f)
-    try:
-      sig.bind(*args, **kwargs)
-    except TypeError:
-      argspec = inspect.getfullargspec(f)
-      n_positional = len(argspec.args)
-      keywords = argspec.kwonlyargs[:len(args) - n_positional]
-      warnings.warn(
-          f"jnp.ndarray.at[...].{f.__name__}: Passing '{keywords[0]}' by position is deprecated. "
-          f"Pass by keyword instead", category=FutureWarning, stacklevel=2)
-      converted_kwargs = dict(unsafe_zip(keywords, args[n_positional:]))
-      return f(*args[:n_positional], **converted_kwargs, **kwargs)
-    else:
-      return f(*args, **kwargs)
-  return wrapped
-
-
 class _IndexUpdateRef:
   """Helper object to call indexed update functions for an (advanced) index.
 
@@ -502,7 +480,6 @@ class _IndexUpdateRef:
   def __repr__(self):
     return f"_IndexUpdateRef({repr(self.array)}, {repr(self.index)})"
 
-  @allow_pass_by_position_with_warning
   def get(self, *, indices_are_sorted=False, unique_indices=False,
           mode=None, fill_value=None):
     """Equivalent to ``x[idx]``.
@@ -519,7 +496,6 @@ class _IndexUpdateRef:
                                      unique_indices=unique_indices, mode=mode,
                                      fill_value=fill_value)
 
-  @allow_pass_by_position_with_warning
   def set(self, values, *, indices_are_sorted=False, unique_indices=False,
           mode=None):
     """Pure equivalent of ``x[idx] = y``.
@@ -533,7 +509,6 @@ class _IndexUpdateRef:
                                    indices_are_sorted=indices_are_sorted,
                                    unique_indices=unique_indices, mode=mode)
 
-  @allow_pass_by_position_with_warning
   def apply(self, func, *, indices_are_sorted=False, unique_indices=False,
             mode=None):
     """Pure equivalent of ``func.at(x, idx)`` for a unary ufunc ``func``.
@@ -557,7 +532,6 @@ class _IndexUpdateRef:
                                    indices_are_sorted=indices_are_sorted,
                                    unique_indices=unique_indices, mode=mode)
 
-  @allow_pass_by_position_with_warning
   def add(self, values, *, indices_are_sorted=False, unique_indices=False,
           mode=None):
     """Pure equivalent of ``x[idx] += y``.
@@ -572,7 +546,6 @@ class _IndexUpdateRef:
                                    indices_are_sorted=indices_are_sorted,
                                    unique_indices=unique_indices, mode=mode)
 
-  @allow_pass_by_position_with_warning
   def multiply(self, values, *, indices_are_sorted=False, unique_indices=False,
                mode=None):
     """Pure equivalent of ``x[idx] *= y``.
@@ -589,7 +562,6 @@ class _IndexUpdateRef:
                                    mode=mode)
   mul = multiply
 
-  @allow_pass_by_position_with_warning
   def divide(self, values, *, indices_are_sorted=False, unique_indices=False,
              mode=None):
     """Pure equivalent of ``x[idx] /= y``.
@@ -606,7 +578,6 @@ class _IndexUpdateRef:
                               indices_are_sorted=indices_are_sorted,
                               unique_indices=unique_indices, mode=mode))
 
-  @allow_pass_by_position_with_warning
   def power(self, values, *, indices_are_sorted=False, unique_indices=False,
             mode=None):
     """Pure equivalent of ``x[idx] **= y``.
@@ -623,7 +594,6 @@ class _IndexUpdateRef:
                               indices_are_sorted=indices_are_sorted,
                               unique_indices=unique_indices, mode=mode))
 
-  @allow_pass_by_position_with_warning
   def min(self, values, *, indices_are_sorted=False, unique_indices=False,
           mode=None):
     """Pure equivalent of ``x[idx] = minimum(x[idx], y)``.
@@ -639,7 +609,6 @@ class _IndexUpdateRef:
                                    indices_are_sorted=indices_are_sorted,
                                    unique_indices=unique_indices, mode=mode)
 
-  @allow_pass_by_position_with_warning
   def max(self, values, *, indices_are_sorted=False, unique_indices=False,
           mode=None):
     """Pure equivalent of ``x[idx] = maximum(x[idx], y)``.